requirements.txt
streamlit==1.28.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
requests==2.31.0
google-generativeai==0.7.2
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0 